# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText
  description: |
    API for WASAText, which is a online messaging application
    similar to whatsapp
  version: 1.0.0
...
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    message:
      description: a single message
      type: object
      properties:
        id:
          type: integer
          description: identifier for message
        status:
          description: status of the message
          type: string
          enum: [ unknown, received, read ]
        messageContent:
          description: the content of the message, string or a file
          type: string
        comment:
          description: emoticon for the message
          type: array
          items:
            type: integer
        timestamp:
          description: time when the message was sent
          type: string
          format: date-time
          example: '2019-10-19 17:12:30'
        sender:
          description: user who sent the message
          type: integer
        conversation:
          description: conversation that the message belongs to
          type: integer
    conversation:
      description: a conversation head
      type: object
      properties:
        id:
          type: integer
          description: identifier for conversation
        snippet:
          description: snippet of the last message, string or picture
          type: string
        name:
          description: name of the sender, username or group name
          type: string
        picture:
          description: picture of the user or group
          type: string
        date:
          description: time when the last message was sent
          type: string
          format: date-time
          example: '2019-10-19 17:12:30'
        conversationContent:
          description: content of the  conversation
          type: array
          items:
            type: integer
        groupchat:
          description: if it is group or not
          type: boolean
        members:
          description: members of the group
          type: array
          items:
            type: integer
...
  parameters:
    userId:
      description: identifier for the user
      name: Id
      in: path
      required: true
      schema:
        type: integer
    conversationId:
      description: identifier for the conversation
      name: conversationId
      in: path
      required: true
      schema:
        type: integer
    messageId:
      description: identifier for the message
      name: messageId
      in: path
      required: true
      schema:
        type: integer
...
paths:
...
  /mainpage/{Id}/conversations:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags:
        - conversations
      summary: gets conversations
      operationId: getMyConversations
      description: list all my conversations
      responses:
        "200":
          description: array of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/conversation"
    post:
      tags:
        - conversations
      summary: starts conversation
      operationId: startNewConversation
      description: start a new conversation with a user by name, or a group
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "201":
          description: conversation successfully created
          content:
            application/json:
              schema:
                type: integer
        "403":
          description: user not found
  /mainpage/{Id}/conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    get:
      tags:
        - conversations
      summary: gets conversation
      operationId: getConversation
      description: get a specific conversation
      responses:
        "200":
          description: object conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/conversation"
        "404":
          description: conversation not found
          content:
            application/json:
              schema:
                type: integer
                enum: [ 0 ]
    post:
      tags:
        - conversations
      summary: sends message
      operationId: sendMessage
      description: send a new message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: message successfully sent
          content:
            application/json:
              schema:
                type: integer
  /mainpage/{Id}/conversations/{conversationId}/conversationsettings/groupname:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    put:
      tags:
        - conversations
      summary: changes group name
      operationId: setGroupName
      description: change the name of the group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "204":
          description: successfully changed the name
        "400":
          description: name could not be changed
  /mainpage/{Id}/conversations/{conversationId}/conversationsettings/grouppicture:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    put:
      tags:
        - conversations
      summary: changes group picture
      operationId: setGroupPhoto
      description: change the picture of the group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "204":
          description: successfully changed the picture
  /mainpage/{Id}/conversations/{conversationId}//{messageId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    delete:
      tags:
        - messages
      summary: deletes message
      operationId: deleteMessage
      description: delete my message
      responses:
        "200":
          description: message successfully deleted
          content:
            application/json:
              schema:
                type: string
        "404":
          description: message not found
    get:
      tags:
        - messages
      summary: gets message
      operationId: getMessage
      description: get a specific message
      responses:
        "200":
          description: object message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        "404":
          description: message not found
          content:
            application/json:
              schema:
                type: integer
                enum: [ 0 ]
    post:
      tags:
        - messages
      summary: forwards message
      operationId: forwardMessage
      description: forward the message to other conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: message successfully forwarded
          content:
            application/json:
              schema:
                type: integer
        "404":
          description: could not find the user
...
  /mainpage/{Id}/conversations/{conversationId}/groupMembers:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    get:
      tags:
        - conversations
      summary: gets group members
      operationId: listGroupMembers
      description: see all the group members
      responses:
        "200":
          description: array of group member names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      tags:
        - conversations
      summary: adds group member
      operationId: addToGroup
      description: add a member to the group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: user successfully added
          content:
            application/json:
              schema:
                type: string
        "404":
          description: user not found
    delete:
      tags:
        - conversations
      summary: leaves group
      operationId: leaveGroup
      description: leave this group
      responses:
        "200":
          description: left the group
          content:
            application/json:
              schema:
                type: string
        "404":
          description: group not found